break dogs out into its own page
