+++ src/main.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App.tsx';
+import DogList from './components/DogList.tsx';
 import './index.css';
 import { Authenticator } from '@aws-amplify/ui-react';
 import { Amplify } from 'aws-amplify';
@@ -9,10 +11,21 @@ import '@aws-amplify/ui-react/styles.css';
 
 Amplify.configure(outputs);
 
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+  },
+  {
+    path: '/dogs',
+    element: <DogList />,
+  },
+]);
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <Authenticator>
-      <App />
+      <RouterProvider router={router} />
     </Authenticator>
   </React.StrictMode>
 );
